post_crud/requirements.txt
pydantic==1.9.1
pymongo==4.3.3
fastapi==0.75.2
uvicorn==0.17.6
python-dotenv==0.20.0
python-multipart==0.0.5
google-cloud-pubsub==2.13.11
google-cloud-storage==2.7.0
bcrypt==4.0.1
passlib==1.7.4
pycryptodome==3.16.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.1
shortuuid==1.0.11
pendulum-2.1.2